Change raids to not depend on channels

+++ Commands/raidattack.js
@@ -1,15 +1,12 @@
 
 module.exports = async function (message, user) {
-    return Promise.all([functions.getObject("mobData", message.channel.id)]).then(ret => {
+    if (user.location == "city") { return functions.replyMessage(message, "You cannot do this in the city!") }
+    return Promise.all([functions.getObject("mobData", user.location)]).then(ret => {
         let raid = ret[0]
         if (raid == false) {
             functions.deleteMessage(message);
-            functions.replyMessage(message, "There is no raid right now!");
+            functions.replyMessage(message, "There is no raid where you are!");
             return;
-        }
-        if (user.dungeonts != undefined) {
-            functions.sendMessage(message.channel, "You cannot do this while in a dungeon!");
-            return
         }
         functions.raidAttack(message, user, raid, "raid")
         functions.setObject("mobData", raid)
